+++ tests/cross_area.rs
@@ -505,6 +505,7 @@ struct FrameWire {
     cursor: Option<CursorWire>,
     hyperlinks: Vec<String>,
     graphics: Vec<u8>,
+    force_full_repaint: bool,
 }
 
 #[allow(dead_code)]